Update AoC11.py
Fixed bug in no_conflict where I was only checking if the floor you were transferring items TOWARD as a legal floor, but I failed to account for if the change you made by moving that item FROM a floor made that prior floor now illegal.  Also added BFS in to compare, and although A* finishes faster by about 20%, BFS explores slightly less states, which is confusing to me.  A* is probably faster despite exploring more states because of some sort of efficiency gain related to popping and appending items using heappush/pop instead of deque?  Not sure.