fixed bug in goal test, as well as successor func
goal test would check distance to remaining numbers, but if no remaining numbers present, program would try to take min of None and crash.  Fixed to return zero if no remaining numbers since the distance to the goal is "zero", and we are at the goal.  Then, this will be popped off frontier first on next pass and goal will be reached.

Also fixed bug in successor function where I was changing the nums_to_visit variable to be equal to the set of the original nums for that state, less a number that an action would have taken us to.  The problem is that this would also modify the nums remaining for any subsequent moves within a particular action sequence, even if those moves did NOT previously cross over a number.  

E.G. I move down one space, and that space contains a number, so I remove that number from the current state's nums_to_visit by rebinding nums_to_visit to be the old state less the number I landed on.  But if the next move in our 4 directional movement did not also land on that same number, it's state would be incorrectly reduced by the number because the prior step modified the nums_to_visit variable.  

To fix this, added "remaining_nums_to_visit" instead of rebinding nums_to_visit, and I set it equal to the current nums_to_visit if it doesn't land on a number, set it to the current nums_to_visit less the current number if it does land on a number.